requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
openpyxl>=3.1.2
python-pptx>=0.6.21
pyarrow>=14.0.0
